requirements.txt
discord.py
pymongo
psycopg2-binary
fastapi
uvicorn